카드 선택 취소 UI 움직임 구현
1.[수정 & 변경사항]
 namespace가 없는 스크립트들 변경
 RatCardUI에서 우측에서 선택한 카드가 튀어 나오도록 제작
 TurnUI와 RatCardUI의 움직임이 확인이 되어
 Input.getkey로 호출하던 문장은 삭제
2.[영향]
 취소 버튼과 선택시 과정에 대해서는 제작하지 않음
 우선 무조건 생길 UI의 움직임만 만들었기에 추후에
 추가, 수정 될 수 있다.
3.[테스트 방법]
