+++ Assets/Art/TestTextures/Old/f72d0d6da0e97e8b.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 417a5a8aecfb81d4599f9c2153173079
+guid: a3dbd76ac5016e04ab8c9e8f8b18c6b0
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
