+++ Assets/Art/TestTextures/Old/fcc991dbab403f87.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cd20f35dd8da8d748b67e9b5bf9d7d22
+guid: 41f0e8a6f0de8904b8e999b9fe7662e0
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
