+++ Assets/Scripts/UI/EventUI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0dd664b306b822d4e9a3a384c7a17ca8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
