+++ Assets/Scripts/UI/EventUI/RatCardUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6c5fb0700896116438cdf3df21ccc147
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
