+++ Assets/Scripts/UI/MainUI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5d24567df7e99d94fb9c9fcba9b44e2f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
